+++ examples/resources/elasticstack_elasticsearch_index/resource.tf
@@ -37,6 +37,10 @@ resource "elasticstack_elasticsearch_index" "my_index" {
       value = "2"
     }
     setting {
+      name  = "index.total_shards_per_node"
+      value = "200"
+    }
+    setting {
       name  = "index.search.idle.after"
       value = "20s"
     }