+++ examples/resources/elasticstack_elasticsearch_index_lifecycle/resource.tf
@@ -27,6 +27,7 @@ resource "elasticstack_elasticsearch_index_lifecycle" "my_ilm" {
         box_type = "hot"
       })
       number_of_replicas = 1
+      total_shards_per_node = 200
     }
   }