+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Add `allow_restricted_indices` setting to security role ([#125](https://github.com/elastic/terraform-provider-elasticstack/issues/125))
 - Add conditional to only set `password` and `password_hash` when a new value is defined ([#127](https://github.com/elastic/terraform-provider-elasticstack/pull/128))
 - Add support for ELASTICSEARCH_INSECURE environment variable as the default of the `insecure` config value ([#127](https://github.com/elastic/terraform-provider-elasticstack/pull/128))
+- Apply `total_shards_per_node` setting in `allocate` action in ILM ([#112](https://github.com/elastic/terraform-provider-elasticstack/issues/112))
 
 ## [0.3.3] - 2023-03-22
 ### Fixed